+++ packages/graphql/src/schema-model/generate-model.test.ts
@@ -81,409 +81,3 @@ describe("ConcreteEntity generation", () => {
         expect(authAnnotation.validatePre).toHaveLength(1);
     });
 });
-
-describe("@authorization directive validation", () => {
-    test("validate pre not array", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { pre: { where: { node: { id: { equals: "$jwt.sub" } } } } }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate pre and post, pre not array", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    validate: {
-                        pre: { where: { node: { id: { equals: "$jwt.sub" } } } }
-                        post: [{ where: { node: { id: { equals: "$jwt.sub" } } } }]
-                    }
-                ) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate pre and post correct", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    validate: {
-                        pre: [{ where: { node: { id: { equals: "$jwt.sub" } } } }]
-                        post: [{ where: { node: { id: { equals: "$jwt.sub" } } } }]
-                    }
-                ) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).not.toThrow();
-    });
-
-    test("validate should be of type object", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: [{ where: { node: { id: { equals: "$jwt.sub" } } } }]) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("filter should be of type object", () => {
-        const typeDefs = gql`
-            type User @authorization(filter: { where: { node: { id: { equals: "$jwt.sub" } } } }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("incorrect directive argument", () => {
-        const typeDefs = gql`
-            type User @authorization(banana: { where: { node: { id: { equals: "$jwt.sub" } } } }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate without correct arguments", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { banana: [{ where: { node: { id: { equals: "$jwt.sub" } } } }] }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate without correct arguments where (2)", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { pre: [{ where: { node: { banana: { equals: "$jwt.sub" } } } }] }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate without correct arguments where (3)", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { pre: [{ where: { node: { id: { thisIsNotRight: "$jwt.sub" } } } }] }) {
-                id: ID!
-                athh: String
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    test("validate without correct arguments where (4)", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { pre: [{ where: { node: { id: { thisIsNotRight: "$jwt.sub" } } } }] }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).toThrow();
-    });
-
-    // correct arguments
-
-    test("validate correct arguments", () => {
-        const typeDefs = gql`
-            type User @authorization(validate: { pre: [{ where: { node: { id: { equals: "valid-string" } } } }] }) {
-                id: ID!
-                name: String!
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(() => generateModel(document)).not.toThrow();
-    });
-
-    // wrong operations
-
-    describe("validate incorrect operations", () => {
-        test("filter operation cannot be CREATE", () => {
-            const typeDefs = gql`
-                type User
-                    @authorization(
-                        filter: [{ operations: [CREATE], where: { node: { id: { equals: "valid-string" } } } }]
-                    ) {
-                    id: ID!
-                    name: String!
-                }
-            `;
-
-            const document = getDocument(typeDefs);
-            expect(() => generateModel(document)).toThrow();
-        });
-
-        test("valid filter operation", () => {
-            const typeDefs = gql`
-                type User
-                    @authorization(
-                        filter: [
-                            {
-                                operations: [READ, CREATE_RELATIONSHIP]
-                                where: { node: { id: { equals: "valid-string" } } }
-                            }
-                        ]
-                    ) {
-                    id: ID!
-                    name: String!
-                }
-            `;
-
-            const document = getDocument(typeDefs);
-            expect(() => generateModel(document)).not.toThrow();
-        });
-    });
-
-    describe("requireAuthentication", () => {
-        test("requireAuthentication should throw if not of type boolean", () => {
-            const typeDefs = gql`
-                type User
-                    @authorization(
-                        filter: [{ requireAuthentication: 22, where: { node: { id: { equals: "valid-string" } } } }]
-                    ) {
-                    id: ID!
-                    name: String!
-                }
-            `;
-
-            const document = getDocument(typeDefs);
-            expect(() => generateModel(document)).toThrow();
-        });
-
-        test("should not throw if undefined", () => {
-            const typeDefs = gql`
-                type User @authorization(filter: [{ where: { node: { id: { equals: "valid-string" } } } }]) {
-                    id: ID!
-                    name: String!
-                }
-            `;
-
-            const document = getDocument(typeDefs);
-            expect(() => generateModel(document)).not.toThrow();
-        });
-    });
-});
-
-/// me
-
-describe("simone suite", () => {
-    // test("1", () => {
-    //     const typeDefs = gql`
-    //         type User
-    //             @authorization(
-    //                 filter: [{ requireAuthentication: true, where: { node: { name: { equals: "1" } } } }]
-    //             ) {
-    //             name: String!
-    //         }
-
-    //     `;
-
-    //     const document = getDocument(typeDefs);
-    //     expect(generateModel(document)).toBeDefined();
-
-    // });
-
-    test.only("node.and ok", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { AND: [{ name: { equals: "1" } }, { bananas: { includes: 1 } }] } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.name.and ok", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { name: { AND: [{ equals: "12" }, { "contains": "2" }] } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.name.and.not ok", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { name: { AND: [{ equals: "12" }, { "contains": "2" }, { NOT: { startsWith: "3" } }] } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.bananas.and ok", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { bananas: { AND: [{ includes: 1 }, { equals: [2, 3] }] } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.bananas.and.and ok", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: {
-                                node: {
-                                    bananas: {
-                                        AND: [
-                                            { includes: 1 }
-                                            { equals: [2, 3] }
-                                            { AND: [{ NOT: { includes: 2 } }, { equals: [5] }] }
-                                        ]
-                                    }
-                                }
-                            }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.bananas.and cannot be combined", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { bananas: { AND: [{ includes: 1 }, { equals: [2, 3] }], includes: 42 } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.name.hocuspocus operator does not exist", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { name: { hocuspocus: [{ equals: "12" }, { "contains": "2" }] } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-
-    test("node.bananas.and.includes.and operator cannot be used here", () => {
-        const typeDefs = gql`
-            type User
-                @authorization(
-                    filter: [
-                        {
-                            requireAuthentication: true
-                            where: { node: { bananas: { AND: [{ includes: { NOT: { 1 } } }, { equals: [2, 3] }] } } }
-                        }
-                    ]
-                ) {
-                name: String!
-                bananas: [Int]
-            }
-        `;
-
-        const document = getDocument(typeDefs);
-        expect(generateModel(document)).toBeDefined();
-    });
-});
